🔨 Refactor: 메인페이지 오늘감정 섹션 분리
