+++ src/app/(after-login)/epigrams/_components/TodayEpigram.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useTodayEpigram } from '@/apis/epigram/epigram.queries';
 import Card from '@/components/Card';
 import { Section } from '@/components/Section';
